recipes-{core,devtools,images}: Add code for SFP persistence
This adds code to rename the sfp ports to sfp0 and sfp1 to match
silkscreen and front-panel by swapping them with a udev rule.

It also adds a systemd oneshot service that will try to migrate
old systemd-networkd configuration files in /data/network that
matched on eth1 and eth2 instead of sfp0 and sfp1.

Finally this service is added to auto-start and disable itself.

Signed-off-by: = Moritz Fischer <mortiz.fischer@ettus.com>