meta-ettus-core: recipes-kernel: Add nixge bugfixes from upstream
Add bugfixes for the nixge network driver that have been queued
upstream. The bug addressed fixes an issue that dereferences an
ERR_PTR leading to an invalid access in the nixge_get_nvmem_address().
Moreover it addresses a memory leak.

These patches are queued upstream and will go away for 4.18 at the
latest.

Signed-off-by: Moritz Fischer <moritz.fischer@ettus.com>