meta-mender-sulfur: conf: machine: Fix MENDER_DEVICE_COMPATIBLE
Fix MENDER_DEVICE_COMPATIBLE such that it reflects all the device types
a given mender image is compatible with are reflected.

The issue at hand is that when built for ni-sulfur-rev6-mender for example
the resulting .mender file would not work with a ni-sulfur-rev5-mender
or ni-sulfur-rev5 device_type device.

Older sulfur devices would ship with filesystems that had different
machine names such as ni-sulfur-rev5, which would not be able to
get upgraded to the new images built for the machines defined in
this layer.

Signed-off-by: Moritz Fischer <moritz.fischer@ettus.com>