meta-mender-e31x: Fix network configuration to match mender
Fix network configuration to match mender upstream changes.
Previously mender used MENDER_DATA_PART and a custom do_deploy
step to install persistent files. This is now deprecated and
is solved by installing persistent files to ${D}/data.

Note: As a result of this we'll only be able to work with the sumo
      version of the meta-mender layer and others.

Signed-off-by: Moritz Fischer <moritz@ettus.com>