contrib: Add build_imgs_package.sh script
This script will build all the files necessary to build a package that
goes onto a fileserver for fetching with the uhd_images_downloader
script. It is equivalent to an FPGA build + calling package_images.py on
the UHD/FPGA side.

Call it like this:

    $ ./meta-ettus/contrib/build_imgs_package.sh $DEV $ARTEFACTNAME

...to build images for device $DEV (e.g., e320, n3xx). It will create
three zip files (one for the sdimg, one for the mender artefact, and one
for the SDK) ready for uploading.

Signed-off-by: Martin Braun <martin.braun@ettus.com>