meta-mender-e31x: systemd: start systemd-networkd after data.mount
When the system is booted with the default configuration for systemd-networkd,
the network configuration cannot be loaded because it consists of symlinks
to /data/network but "data" is not mounted at this point in time.

Change the systemd-network service to wait for data.mount to be available
before starting this service.

Signed-off-by: Joerg Hofrichter <joerg.hofrichter@ni.com>