meta-sulfur: ec firmware: generate symlinks up to rev. 10
This prepares the updated embedded controller (ec) firmware
for the sulfur rev. 10 hardware.

The same firmware code will be used for all revisions of
hardware (starting from rev. 4) because the behavior of
the firmware changes dynamically on runtime based on the
revision field in the eeprom.

For Rev5+, the bootdelay timeout is increased.

For Rev10+, additional GPIO pins for enabling the 12V
for the daughter boards and for enabling the fans
will be used.

Signed-off-by: Joerg Hofrichter <joerg.hofrichter@ni.com>