meta-[mender-]sulfur: added comments about compat. of rev.10
Generated dummy machine configuration ni-sulfur-rev10 which
internally includes ni-sulfur-rev6.

Added comment that the rev. 10 is compatible to rev. 6 with
regards to the device tree. If the only change is the embedded
controller (ec) firmware, it is sufficient to provide a new
ec firmware and increase the "mcu_compat" field in eeprom
accordingly. The "dt_compat" field can stay unchanged.

Signed-off-by: Joerg Hofrichter <joerg.hofrichter@ni.com>