meta-ettus-core: recipes-kernel: linux-yocto: fixed-link for nixge
Backported changes in mainline to nixge / fixed-phy ethernet drivers which
support setting the PHY link state based on a GPIO.

In case of NI USRP devices, the GPIOs are provided from the FPGA (when
using the approriate FPGA image). The change is necessary because reading
the link detection state from MDIO is unreliable for 10 Gbe.

The device tree needs to be changed in order to use the link state from
GPIO, see Documentation/devicetree/bindings/net/nixge.txt. Also, the FPGA
image needs to provide the necessary GPIO lines.

Signed-off-by: Joerg Hofrichter <joerg.hofrichter@ni.com>