meta-{ettus-core,neon,sulfur}: linux-yocto: move CONFIG_CROS_EC_THERMAL
The configuration CONFIG_CROS_EC_THERMAL is introduced with the kernel patches
in meta-sulfur and meta-neon. It is not available for e31x, so enabling this
configuration for this device leads to a warning of CONFIG_CROS_EC_THERMAL
being unknown. The solution is to move this configuration to the machine
dependent configuration (sulfur.cfg, neon.cfg).

Signed-off-by: Joerg Hofrichter <joerg.hofrichter@ni.com>