meta-ettus-core: contrib: Fix output path in build_imgs_package.sh
The output path of the SDKs and images depends on the libc version used.
We used to use glibc, and the `tmp-glibc` output path was hard coded
into the script (for musl, it's `tmp-musl`). Since we've now permanently
switched to musl, we change that to tmp-musl, but we use a variable like
we should instead of hard-coding it.

Signed-off-by: Martin Braun <martin.braun@ettus.com>