meta-neon: u-boot: fix neon_read_ethaddr for neon-rev2
When the changes for rev2 were made, neon_read_ethaddr() didn't have the sequence
number of the i2c tunnel updated. neon_probe_mboard() was updated and works fine.
This is only a problem if you want to use ethernet in u-boot.

Reported & patch provided by: aholtzma (on github.com)
Signed-off-by: Joerg Hofrichter <joerg.hofrichter@ni.com>