meta-ettus: spicadence Revert cs gpio descriptors
Newer versions of the driver introduced a change to centrally handle
gpio for chipselect. This change broke SPI on embedded USRP devices.
Revisit removing patch in the future when the code churn stabilizes.
This also removes changes that were intended to fix the original
patch. This patch reverts spi-cadence.c from v5.2 to v5.0.

Signed-off-by: steven.koo <steven.koo@ni.com>