meta-ettus: Remove PHY_FORCING state in phy state
nixge's 10G phy is a clause 45 phy to support 10G speeds. PHY_FORCING's
genphy_update_link isn't aware of C45 phys and calls the C22
implementation, which causes an error on the nixge's phy. PHY_FORCING is
removed in future versions of the kernel because it was used to support
an old use case which doesn't apply anymore. This upstream patch fixes
the phy error.

Signed-off-by: Steven Koo <steven.koo@ni.com>