meta-ettus-core: classes: add run_tests class
This class adds the task "run_tests" which runs ctest in the build
directory. The results are written to the directory specified by
TESTRESULTS_DIR in a subdirectory named after the recipe.

The default of TESTRESULTS_DIR is set to ${WORKDIR}/testresults.
You can overwrite this in your conf/local.conf file.

Signed-off-by: Joerg Hofrichter <joerg.hofrichter@ni.com>