meta-ettus: set e310-sg3 clock to 867 MHz
sg1 and sg3 should have different cpu clock speeds. They should be 667
Mhz and 867 MHz respectively. The clock speed is controlled by a clock
selector and divider (0XF8000120, 0x1F003F30U) and a clock multiplier
(0XF8000100, 0x0007F000U).

The existing implementation sets both the sg1 and sg3 registers to the
to the same values:
ARM PLL (33.33Mhz) * Multiplier (0x28/40) / Divider (0x2/2).
This is a resulting clock of 667 MHz.

This change modifies the sg3 multiplier to 0x34/52 so the clock becomes:
ARM PLL (33.33Mhz) * Multiplier (0x34/52) / Divider (0x2/2).
This is a resulting clock of 867 MHz.

Signed-off-by: Steven Koo <steven.koo@ni.com>