ci: don't delete directories cached by pipeline
The "save" stage of the pipeline cache runs after all steps in the job,
so deleting these directories here will cause that save step to fail,
and the artifacts not be cached.

So, if these directories are meant to be cached, leave them as-is. Still
remove the build directory, which is where the bulk of the space is...