contrib: create_packages: compress the sdimg
Compress the sdimg with bzip2 before including it in the zipfile. This
might seem a bit redundant, but it keeps the extracted size small
(whereas right now, extracting this zipfile will require at least as
much space as the SD card / eMMC on the device).

bmaptool supports flashing images that have been bzip2 compressed, and
users of other utilities can extract the .sdimg.bz2 to get the raw
.sdimg

Signed-off-by: Michael Auchter <michael.auchter@ni.com>