sulfur: Update firmware to resolve apshutdown
When using apshutdown from the SCU, the watchdog for the Linux OS will
trip. This change points to firmware that will ignore the watchdog if
the the device is in the S3 or lower power state. apshutdown is useful
for controlling the device programatically to get it in a known off
state. This can be used for test systems.

Signed-off-by: Steven Koo <steven.koo@ni.com>