meta-ettus-bsp: build fitImage-manufacturing for all targets
This builds another fitImage called fitImage-manufacturing-image
which includes an initramfs image which is based on the
manufacturing-image.

Kernel + initramfs live completely in RAM which means that
there is no partition of the SD card mounted. This makes it
possible to flash the SD card with a complete filesystem
image. The manufacturing-image provides the tool bmap-tool for
flashing.

Hint on CI specifics: As part of the pipeline step "Aggregate
images", the file fitImage-manufacturing-image is renamed to
fitImage-manufacturing.

Signed-off-by: Jörg Hofrichter <joerg.hofrichter@emerson.com>