ci: Add step template to use NI internal sbom management tools
Create new step template steps-sbom-convert-spdx.yml to
- Convert yocto produced SPDX sbom to CycloneDX format
- Archive additional documents as additional pipleine artifact
- Trigger common NI pipeline to upload SBOM to internal tracking tool

Template is using a custom python environment and uses a private
python package to validate all detected SPDX SBOMs and convert them
to a CycloneDX format, which NI is using internally to manage, track
and analyze SBOM content information.
Step will attempt to parse the file system build manifest and derive
a 4 field numeric version belonging to the uhd source component.
Failure to determine such a version generates an error and all steps
are skipped. The manifest parsing is implemented as a python script.